Adiciona estilo para filtro do de prestadores de serviço e adiciona campos que faltavam para visualizar as oficinas

+++ PROJETO/src/views/main-page/Oficina/perfil-oficina.php
@@ -14,8 +14,24 @@ $oficina_id = $_SESSION['id_oficina'];
 
 // verifica se o form foi enviado via POST
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    // verifica se o botão de excluir foi pressionado
-    if (isset($_POST['excluir_perfil']) && $_POST['excluir_perfil'] === '1') {
+     // verifica se o botão de excluir foi pressionado
+     if (isset($_POST['excluir_perfil']) && $_POST['excluir_perfil'] === '1') {
+        
+        // Verifica se existem funcionários cadastrados
+        $sqlCheckFuncionarios = "SELECT COUNT(*) as total FROM funcionarios WHERE id_oficina = ?";
+        $stmtCheck = $conexao->prepare($sqlCheckFuncionarios);
+        $stmtCheck->bind_param("i", $oficina_id);
+        $stmtCheck->execute();
+        $resultCheck = $stmtCheck->get_result();
+        $row = $resultCheck->fetch_assoc();
+        $totalFuncionarios = $row['total'];
+        $stmtCheck->close();
+        
+        if ($totalFuncionarios > 0) {
+            // Redireciona para a página de funcionários se houver funcionários cadastrados
+            echo "<script>alert('Você não pode excluir a oficina enquanto houver funcionários cadastrados. Por favor, remova todos os funcionários primeiro.'); window.location.href='/fixTime/PROJETO/src/views/main-page/Oficina/funcionarios.php';</script>";
+            exit();
+        }
 
         $sqlDelete = "DELETE FROM oficina WHERE id_oficina = ?";
         $stmtDelete = $conexao->prepare($sqlDelete);
@@ -134,6 +150,16 @@ $conexao->close();
 
                 </a>
                 <ul class="space-y-2 font-medium">
+
+                    <li>
+                        <a href="/fixTime/PROJETO/src/views/main-page/Oficina/funcionarios.php" class="flex items-center p-2 text-gray-900 rounded-lg  hover:bg-gray-100  group">
+                        <svg class="shrink-0 w-6 h-6 text-gray-500 transition duration-75 group-hover:text-gray-900" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
+                            <path stroke="currentColor" stroke-linecap="round" stroke-width="2" d="M4.5 17H4a1 1 0 0 1-1-1 3 3 0 0 1 3-3h1m0-3.05A2.5 2.5 0 1 1 9 5.5M19.5 17h.5a1 1 0 0 0 1-1 3 3 0 0 0-3-3h-1m0-3.05a2.5 2.5 0 1 0-2-4.45m.5 13.5h-7a1 1 0 0 1-1-1 3 3 0 0 1 3-3h3a3 3 0 0 1 3 3 1 1 0 0 1-1 1Zm-1-9.5a2.5 2.5 0 1 1-5 0 2.5 2.5 0 0 1 5 0Z"/>
+                        </svg>
+
+                            <span class="flex-1 ms-3 whitespace-nowrap">Meus Funcionarios</span>
+                        </a>
+                    </li>
 
                     <li>
                         <a href="/fixTime/PROJETO/src/views/main-page/Oficina/perfil-oficina.php" class="flex items-center p-2 text-gray-900 rounded-lg  hover:bg-gray-100  group">
